### 2023-01-06 #5049 @editorialbot
[PRE REVIEW]: A C++ Implementation of a Cartesian Impedance Controller for Robotic Manipulators
**Submitting author:** @matthias-mayr (<a href="http://orcid.org/0000-0002-8198-3154">Matthias Mayr</a>)
**Repository:** https://github.com/matthias-mayr/Cartesian-Impedance-Controller
**Branch with paper.md** (empty if default branch): joss_paper
**Version:** 1.0.0
