### 2023-01-06 #5049 @kyleniemeyer
@matthias-mayr thanks for your submission to JOSS! We will use this issue to find an editor and reviewers for your paper.

Although the paper compiled, it looks like there was at least one issue with the bibliography. Could you clean up the `month` field in the ROS entry to just be `month = may` ?
