+++ res/paper.bib
@@ -89,8 +89,8 @@
   number={3.2},
   year={2009},
   address = "Kobe, Japan",
-  %note = {May 12--17}
-  month = may # {~12} # {--} # {17,}
+  %note = {May 12--17},
+  month = "May" # {~12} # {--} # {17,}
 }
 
 @inproceedings{lawrence:1988,
