### 2023-01-06 #5049 @matthias-mayr
> could you please suggest 5-6 reviewers from this list that you consider appropriate to vet this paper? While selecting, you should confirm that the domain expertise of the reviewer matches the subject matter of your paper and software.

@adi3: If the selection happens purely based on the information in the sheet, I found reviewers for robotics & automatic control with C++ knowledge: jizecn, vmatos, destogl, sea-bass, JHartzer, askuric
