### 2023-01-26 104ec2e
Update: Damping factor clarification and limits
- Clarifies the use of the damping factors
- Allows a new limit of 2, which is applied on top of 2*sqrt(stiffness)
