### 2023-03-01 #5194 @JHartzer
I am able to compile the code, but there don't appear to be unit/integration tests that I can run to verify functionality. I am unsure to what extent automated tests could be used to test the ROS functionality, but they could prove useful to validate the remaining core functions of `cartesian_impedance_controller.cpp`. I don't think think this is an acceptance-blocker.
