### 2023-03-04 #5194 @matthias-mayr
Thanks for the quick reaction. In general, loading controllers in ROS is not the easiest thing. It might make sense to include a simple mock-up example like this one in the [cartesian_controllers repository](https://github.com/fzi-forschungszentrum-informatik/cartesian_controllers/tree/master/cartesian_controller_examples) into this repo as well.
