### 2023-03-08 #5194 @JHartzer
I think that would improve the quality of the code base, but as it stands I believe the README describes the code use sufficiently for acceptance of the paper.
