### 2023-05-15 #5194 @matthias-mayr
Sure @adi3. Based on C++ knowledge & "robotics" as well as "control systems" appearing the covered domains, the suggested candidates are:
> destogl, shravantata, zamantousif, sea-bass, vmatos, jizecn

We also marked all except `shravantata` 3 months ago. Hopefully their availability has changed.
I leave it up to you to mark them.
