### 2023-06-28 d2a4cdd
Update: Paper references and acknowledgement
