+++ res/paper.bib
@@ -8,6 +8,17 @@
   pages={1485-1492},
   doi={10.1109/CASE49997.2022.9926713}}
 
+@INPROCEEDINGS{mayr23wiping,
+  author={Mayr, Matthias and Ahmad, Faseeh and Dürr, Alexander and Krueger, Volker},
+  booktitle={2023 IEEE 19th International Conference on Automation Science and Engineering (CASE)},
+  title={Using Knowledge Representation and Task Planning for Robot-agnostic Skills on the Example of Contact-Rich Wiping Tasks},
+  year={2023},
+  volume={},
+  number={},
+  pages={},
+  doi={}}
+
+
 @misc{mayr22skireil,
   author = {Mayr, Matthias and Ahmad, Faseeh and Chatzilygeroudis, Konstantinos and Nardi, Luigi and Krueger, Volker},
   title = {Skill-based Multi-objective Reinforcement Learning of Industrial Robot Tasks with Planning and Knowledge Integration},
@@ -34,6 +45,16 @@
   booktitle={2022 IJCAI Planning and Reinforcement Learning Workshop},
   year={2022}
 }
+
+@INPROCEEDINGS{ahmad23generalization,
+  author={Ahmad, Faseeh and Mayr, Matthias and Krueger, Volker},
+  booktitle={IEEE/RSJ International Conference on Intelligent Robots and Systems (IROS)},
+  title={Learning to Adapt the Parameters of Behavior Trees and Motion Generators to Task Variations},
+  year={2023},
+  volume={},
+  number={},
+  pages={},
+  doi={}}
 
 @book{springer:2016,
   title={Springer Handbook of Robotics},
