### 2023-08-11 #5194 @ShravanTata
Hello, Sorry about my delay in reviewing the work. Excellent work! Here are some of my feedback:

- I have tested the library in the following environment
  - M1 Mac with MacOS Ventura 13.5
  - Docker: ros:noetic-robot
- *Installation:* Does installation proceed as outlined in the documentation?
  + Add yaml-cpp as a dependency
  + Mention which ROS distros are supported
  + Add catkin_tools as dependency src_sh{apt install python3-catkin-tools}
  + Some of the above might be standard in a ROS environment but always good to include more.
- *Automated tests:* Are there automated tests or manual steps described so that the functionality of the software can be verified?
   - I cannot confirm as there are no examples provided or documentation to do it. I agree with @JHartzer to have some sort of testing included.
- *Example usage:* Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).
   + Same as above. Authors do claim that it has been tested in real hardware and simulation though.
