### 2023-08-30 #5194 @matthias-mayr
Thanks a lot for the review and the comments.

> Installation: Does installation proceed as outlined in the documentation?
> * Add yaml-cpp as a dependency

Good catch. In a minimal ROS installation it's not there. The same is the case for eigen3 and python-rosdep. It works in the CI pipeline snce the OSRF ROS image seems to have more packages installed.

> * Mention which ROS distros are supported

Will do
> * Add catkin_tools as dependency src_sh{apt install python3-catkin-tools}

I am a bit hesitant to do that since it's only a dependency of the readme, which is why I wrote `catkin build # or catkin_make` there. I can reverse the order though, since `catkin_make` is covered by a standard ROS installation.

> Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?
> * I cannot confirm as there are no examples provided or documentation to do it. I agree with @JHartzer to have some sort of testing included.

We have a CI pipeline that automatically builds the whole repository on the latest ROS noetic docker container. As said above, some more detailed tests can of course be added.

> Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).
> * Same as above. Authors do claim that it has been tested in real hardware and simulation though.

We included some explanation on how to use it with two popular robot systems in the Readme.<br>
However, I agree that we could also supply a minimal launch file with a fake robot description in the same way that the [cartesian_controllers repository](https://github.com/fzi-forschungszentrum-informatik/cartesian_controllers/blob/master/cartesian_controller_examples/launch/examples.launch) does.

It would take a bit of time until I can allocate a time block to address these. I can comment here once it's done.
