+++ res/paper.bib
@@ -9,25 +9,25 @@
   doi={10.1109/CASE49997.2022.9926713}}
 
 @INPROCEEDINGS{mayr23wiping,
-  author={Mayr, Matthias and Ahmad, Faseeh and Dürr, Alexander and Krueger, Volker},
+  author={Mayr, Matthias and Ahmad, Faseeh and Duerr, Alexander and Krueger, Volker},
   booktitle={2023 IEEE 19th International Conference on Automation Science and Engineering (CASE)},
   title={Using Knowledge Representation and Task Planning for Robot-agnostic Skills on the Example of Contact-Rich Wiping Tasks},
   year={2023},
   volume={},
   number={},
-  pages={},
-  doi={}}
+  pages={1-7},
+  doi={10.1109/CASE56687.2023.10260413}}
 
 
-@misc{mayr22skireil,
+@INPROCEEDINGS{mayr22skireil,
   author = {Mayr, Matthias and Ahmad, Faseeh and Chatzilygeroudis, Konstantinos and Nardi, Luigi and Krueger, Volker},
-  title = {Skill-based Multi-objective Reinforcement Learning of Industrial Robot Tasks with Planning and Knowledge Integration},
-  year = {2022},  
-  url = {https://arxiv.org/abs/2203.10033},
-  keywords = {Robotics (cs.RO), Machine Learning (cs.LG), FOS: Computer and information sciences, FOS: Computer and information sciences},
-  publisher = {arXiv},
-  copyright = {Creative Commons Attribution Share Alike 4.0 International},
-  doi = {10.48550/ARXIV.2203.10033}
+  booktitle={2022 IEEE International Conference on Robotics and Biomimetics (ROBIO)},
+  title={Skill-based Multi-objective Reinforcement Learning of Industrial Robot Tasks with Planning and Knowledge Integration},
+  year={2022},
+  volume={},
+  number={},
+  pages={1995-2002},
+  doi={10.1109/ROBIO55434.2022.10011996}
 }
 
 @InProceedings{FDCC,
