### 2023-10-27 327ac0a
Throw exception for a negative damping.
