+++ test/test2_ros.cpp
@@ -1,0 +1,162 @@
+#include <Eigen/Dense>
+#include <gtest/gtest.h>
+#include <ros/ros.h>
+#include <geometry_msgs/WrenchStamped.h>
+#include <geometry_msgs/PoseStamped.h>
+#include <std_msgs/Float64MultiArray.h>
+#include <dynamic_reconfigure/Config.h>
+
+
+//ROSBaseTests tested that key topics in the library are being written.
+TEST(ROSBaseTests, referenceposeTests)
+{
+    ros::NodeHandle n;
+    ros::Duration timeout(10);
+  for (int i = 0; i < 5; i++) {
+    auto msg_pose = ros::topic::waitForMessage<geometry_msgs::PoseStamped>(
+        "/CartesianImpedance_trajectory_controller/reference_pose", n, timeout);
+
+    ASSERT_TRUE(msg_pose != nullptr) << "Did not receive message on iteration " << i + 1;
+
+    EXPECT_LE(std::abs(msg_pose->pose.orientation.x), 11);
+  }
+
+}
+
+TEST(ROSBaseTests, commandedtorqueTests)
+{
+    ros::NodeHandle n;
+    ros::Duration timeout(5);
+  for (int i = 0; i < 5; i++) {
+    auto msg_torque = ros::topic::waitForMessage<std_msgs::Float64MultiArray>(
+        "/CartesianImpedance_trajectory_controller/commanded_torques", n, timeout);
+
+    ASSERT_TRUE(msg_torque != nullptr) << "Did not receive message on iteration " << i + 1;
+
+    EXPECT_LE(std::abs(msg_torque->data[0]), 1000.0);
+  }
+
+}
+
+//ROSReconfigureTests tested that we are able to modify external wrench, virtual stiffness, and damping online.
+TEST(ROSReconfigureTests, wrenchTest) {
+    ros::NodeHandle n;
+    ros::Duration timeout(10);
+  for (int i = 0; i < 5; i++) {
+    auto msg_wup = ros::topic::waitForMessage<dynamic_reconfigure::Config>(
+        "/CartesianImpedance_trajectory_controller/cartesian_wrench_reconfigure/parameter_updates", n, timeout);
+
+    ASSERT_TRUE(msg_wup != nullptr) << "Did not receive message on iteration " << i + 1;
+
+    for(const dynamic_reconfigure::DoubleParameter& double_param : msg_wup->doubles) {
+        if(double_param.name == "f_x") {
+            EXPECT_EQ(double_param.value, 5.0);
+            break; 
+        }
+    }
+  }   
+}
+
+TEST(ROSReconfigureTests, stiffnessTest) {
+    ros::NodeHandle n;
+    ros::Duration timeout(5);
+  for (int i = 0; i < 5; i++) {
+    auto msg_stup = ros::topic::waitForMessage<dynamic_reconfigure::Config>(
+        "/CartesianImpedance_trajectory_controller/stiffness_reconfigure/parameter_updates", n, timeout);
+
+    ASSERT_TRUE(msg_stup != nullptr) << "Did not receive message on iteration " << i + 1;
+
+    for(const dynamic_reconfigure::DoubleParameter& double_param : msg_stup->doubles) {
+        if(double_param.name == "translation_x") {
+            EXPECT_EQ(double_param.value, 100.0);
+            break; 
+        }
+    }
+  }   
+}
+
+TEST(ROSReconfigureTests, dampingTest) {
+    ros::NodeHandle n;
+    ros::Duration timeout(5);
+  for (int i = 0; i < 5; i++) {
+    auto msg_dup = ros::topic::waitForMessage<dynamic_reconfigure::Config>(
+        "/CartesianImpedance_trajectory_controller/damping_factors_reconfigure/parameter_updates", n, timeout);
+
[…]
