### 2023-10-31 49e145d
Tests: Improves base library tests
Also defines more consistent behavior in the library
