+++ test/base_tests.cpp
@@ -1,44 +1,45 @@
 #include "cartesian_impedance_controller/cartesian_impedance_controller.h"
 #include <Eigen/Dense>
 #include <gtest/gtest.h>
-#include <ros/ros.h>
+
 
 using namespace cartesian_impedance_controller;
 
 
 
-CartesianImpedanceController cont; // Created an instance of the controller
+CartesianImpedanceController cont; // Create an instance of the controller
 
 
 TEST(ControllerTests, initializationTests){
-	Eigen::Vector3d position_d_target(1.0, 2.0, 3.0);
-    	Eigen::Quaterniond orientation_d_target(1.0, 0.0, 0.0, 0.0);
- 
-	EXPECT_NO_THROW(cont.initDesiredPose(position_d_target, orientation_d_target)); 
-	
-	
-	Eigen::VectorXd q_d_nullspace_target(7);
-    	q_d_nullspace_target << 1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0;
-	int n = 7;
-	cont.setNumberOfJoints(n);
-	EXPECT_NO_THROW(cont.initNullspaceConfig(q_d_nullspace_target)); //Should not fail since we have the same number of joints as q_d_nullspace_target elements
-	
-	cont.setNumberOfJoints(n+1);
-	EXPECT_DEBUG_DEATH(cont.initNullspaceConfig(q_d_nullspace_target), "Nullspace target needs to same size as n_joints_");
-	
+	const Eigen::Vector3d position_d_target(1.0, 2.0, 3.0);
+    const Eigen::Quaterniond orientation_d_target(1.0, 0.0, 0.0, 0.0);
+
+	EXPECT_NO_THROW(cont.initDesiredPose(position_d_target, orientation_d_target));
 }
 
+TEST(ControllerTests, nullspaceConfigTests){
+	Eigen::VectorXd q_d_nullspace_target(7);
+    q_d_nullspace_target << 1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0;
+	const int n = 7;
+	cont.setNumberOfJoints(n);
+	EXPECT_NO_THROW(cont.initNullspaceConfig(q_d_nullspace_target)); //Should not fail since we have the same number of joints as q_d_nullspace_target elements
+
+	cont.setNumberOfJoints(n+1);
+	EXPECT_DEBUG_DEATH(cont.initNullspaceConfig(q_d_nullspace_target), "Nullspace target needs to same size as n_joints_"); //Should fail since we have differences between the number of joints and the number of q_d_nullspace_target elements
+}
 
 TEST(ControllerTests, setjointnumberTests){
-	int n = 7; 
+	const int n = 7;
 	EXPECT_NO_THROW(cont.setNumberOfJoints(n)); //This value of n shouldn't throw an exception.
 	EXPECT_ANY_THROW(cont.setNumberOfJoints(-n)); //For a negative number of joints, we expect to catch the error.
 }
 
 TEST(ControllerTests, dampingTests){
-	double d = 3.2;
+	const double d = 3.2;
 	EXPECT_NO_THROW(cont.setDampingFactors(d,d,d,d,d,d,d)); //This should work.
-	EXPECT_ANY_THROW(cont.setDampingFactors(-d,d,d,d,d,d,d)); //This should not work since damping should be positive.
+	EXPECT_NO_THROW(cont.setDampingFactors(-d,d,d,d,d,d,d)); //Negative damping factors are ignored outside debug builds
+
[…]
