+++ test/base_tests_old.cpp
@@ -1,124 +1,0 @@
-#include "cartesian_impedance_controller/cartesian_impedance_controller.h"
-#include <Eigen/Dense>
-#include <gtest/gtest.h>
-
-
-using namespace cartesian_impedance_controller;
-
-
-
-CartesianImpedanceController cont; // Create an instance of the controller
-
-
-TEST(ControllerTests, initializationTests){
-	Eigen::Vector3d position_d_target(1.0, 2.0, 3.0);
-    	Eigen::Quaterniond orientation_d_target(1.0, 0.0, 0.0, 0.0);
- 
-	EXPECT_NO_THROW(cont.initDesiredPose(position_d_target, orientation_d_target)); 
-	
-	
-	Eigen::VectorXd q_d_nullspace_target(7);
-    	q_d_nullspace_target << 1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0;
-	int n = 7;
-	cont.setNumberOfJoints(n);
-	EXPECT_NO_THROW(cont.initNullspaceConfig(q_d_nullspace_target)); //Should not fail since we have the same number of joints as q_d_nullspace_target elements
-	
-	cont.setNumberOfJoints(n+1);
-	EXPECT_DEBUG_DEATH(cont.initNullspaceConfig(q_d_nullspace_target), "Nullspace target needs to same size as n_joints_"); //Should fail since we have differences between the number of joints and the number of q_d_nullspace_target elements
-	
-}
-
-
-TEST(ControllerTests, setjointnumberTests){
-	int n = 7; 
-	EXPECT_NO_THROW(cont.setNumberOfJoints(n)); //This value of n shouldn't throw an exception.
-	EXPECT_ANY_THROW(cont.setNumberOfJoints(-n)); //For a negative number of joints, we expect to catch the error.
-}
-
-TEST(ControllerTests, dampingTests){
-	double d = 3.2;
-	EXPECT_NO_THROW(cont.setDampingFactors(d,d,d,d,d,d,d)); //This should work.
-	EXPECT_ANY_THROW(cont.setDampingFactors(-d,d,d,d,d,d,d)); //This should not work since damping should be positive.
-}
-
-TEST(ControllerTests, stiffnessTests) {
-    Eigen::Matrix<double, 7, 1> stiffness_good, stiffness_bad;
-    stiffness_good << 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0;
-    stiffness_bad << -1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0;
-    double t_x = 1.0, t_y = 1.0, t_z = 1.0, r_x = 1.0, r_y = 1.0, r_z = 1.0, n = 1.0;
-    
-    EXPECT_NO_THROW(cont.setStiffness(stiffness_good, false));
-    EXPECT_DEBUG_DEATH(cont.setStiffness(stiffness_bad, false), "Stiffness values need to be positive.");
-    EXPECT_NO_THROW(cont.setStiffness(t_x, t_y, t_z, r_x, r_y, r_z, false));
-    EXPECT_NO_THROW(cont.setStiffness(t_x, t_y, t_z, r_x, r_y, r_z, n, false));
-    
-    EXPECT_NO_THROW(cont.setStiffness(stiffness_good, true));
-    EXPECT_DEBUG_DEATH(cont.setStiffness(stiffness_bad, true), "Stiffness values need to be positive.");
[…]
