### 2023-10-31 50b9e34
New: Clean up ROS tests, switch to example robot
- Minor modifications to the CMakeLists
- Bumps package version to 1.1
- Switches to the example controller configuration
