+++ test/ros_tests.cpp
@@ -6,15 +6,16 @@
 #include <std_msgs/Float64MultiArray.h>
 #include <dynamic_reconfigure/Config.h>
 
+const std::string ctrl_name = "/CartesianImpedance_trajectory_controller";
 
 //ROSBaseTests tested that key topics in the library are being written.
 TEST(ROSBaseTests, referenceposeTests)
 {
-    ros::NodeHandle n;
-    ros::Duration timeout(10);
+  ros::NodeHandle n;
+  ros::Duration timeout(1);
   for (int i = 0; i < 5; i++) {
     auto msg_pose = ros::topic::waitForMessage<geometry_msgs::PoseStamped>(
-        "/CartesianImpedance_trajectory_controller/reference_pose", n, timeout);
+        ctrl_name + "/reference_pose", n, timeout);
 
     ASSERT_TRUE(msg_pose != nullptr) << "Did not receive message on iteration " << i + 1;
 
@@ -25,14 +26,13 @@ TEST(ROSBaseTests, referenceposeTests)
 
 TEST(ROSBaseTests, commandedtorqueTests)
 {
-    ros::NodeHandle n;
-    ros::Duration timeout(5);
+  ros::NodeHandle n;
+  ros::Duration timeout(5);
+  const std::string torque_topic {ctrl_name + "/commanded_torques"};
   for (int i = 0; i < 5; i++) {
-    auto msg_torque = ros::topic::waitForMessage<std_msgs::Float64MultiArray>(
-        "/CartesianImpedance_trajectory_controller/commanded_torques", n, timeout);
+    auto msg_torque = ros::topic::waitForMessage<std_msgs::Float64MultiArray>(torque_topic, n, timeout);
 
     ASSERT_TRUE(msg_torque != nullptr) << "Did not receive message on iteration " << i + 1;
-
     EXPECT_LE(std::abs(msg_torque->data[0]), 1000.0);
   }
 
@@ -40,11 +40,11 @@ TEST(ROSBaseTests, commandedtorqueTests)
 
 //ROSReconfigureTests tested that we are able to modify external wrench, virtual stiffness, and damping online.
 TEST(ROSReconfigureTests, wrenchTest) {
-    ros::NodeHandle n;
-    ros::Duration timeout(10);
+  ros::NodeHandle n;
+  ros::Duration timeout(10);
   for (int i = 0; i < 5; i++) {
     auto msg_wup = ros::topic::waitForMessage<dynamic_reconfigure::Config>(
-        "/CartesianImpedance_trajectory_controller/cartesian_wrench_reconfigure/parameter_updates", n, timeout);
+        ctrl_name + "/cartesian_wrench_reconfigure/parameter_updates", n, timeout);
 
     ASSERT_TRUE(msg_wup != nullptr) << "Did not receive message on iteration " << i + 1;
 
@@ -58,11 +58,11 @@ TEST(ROSReconfigureTests, wrenchTest) {
 }
 
 TEST(ROSReconfigureTests, stiffnessTest) {
-    ros::NodeHandle n;
-    ros::Duration timeout(5);
+  ros::NodeHandle n;
+  ros::Duration timeout(5);
   for (int i = 0; i < 5; i++) {
     auto msg_stup = ros::topic::waitForMessage<dynamic_reconfigure::Config>(
-        "/CartesianImpedance_trajectory_controller/stiffness_reconfigure/parameter_updates", n, timeout);
+        ctrl_name + "/stiffness_reconfigure/parameter_updates", n, timeout);
 
     ASSERT_TRUE(msg_stup != nullptr) << "Did not receive message on iteration " << i + 1;
 
@@ -80,7 +80,7 @@ TEST(ROSReconfigureTests, dampingTest) {
     ros::Duration timeout(5);
   for (int i = 0; i < 5; i++) {
[…]
