+++ CMakeLists.txt
@@ -1,4 +1,5 @@
 cmake_minimum_required(VERSION 2.8.3)
+cmake_policy(SET CMP0048 NEW) # Version not in project() command
 project(cartesian_impedance_controller)
 
 find_package(Boost 1.49 REQUIRED)
@@ -100,22 +101,15 @@ install(DIRECTORY include/${PROJECT_NAME}/
   PATTERN ".svn" EXCLUDE)
 
 ## Testing
-#Cpp tests
 if (CATKIN_ENABLE_TESTING)
+  # Base library tests
 	find_package(rostest REQUIRED)
 	add_rostest_gtest(base_tests test/base.test test/base_tests.cpp)
-	if(TARGET base_tests)
-	    target_link_libraries(base_tests ${PROJECT_NAME} ${Eigen3_LIBRARIES} ${catkin_LIBRARIES})
-	endif()
-endif()
+  target_link_libraries(base_tests ${PROJECT_NAME} ${Eigen3_LIBRARIES} ${catkin_LIBRARIES})
 
-#ROS tests
-if (CATKIN_ENABLE_TESTING)
-	find_package(rostest REQUIRED)
+  # ROS tests
 	add_rostest_gtest(ros_tests test/ros.test test/ros_tests.cpp)
-	if(TARGET ros_tests)
-	    target_link_libraries(ros_tests ${PROJECT_NAME} ${Eigen3_LIBRARIES} ${catkin_LIBRARIES})
-	endif()
+  target_link_libraries(ros_tests ${PROJECT_NAME} ${Eigen3_LIBRARIES} ${catkin_LIBRARIES})
 endif()
 
 
