+++ res/config/example_controller.yaml
@@ -1,3 +1,7 @@
+# An example configuration that works with the example
+# 
+# It is also used in the integration tests
+
 CartesianImpedance_trajectory_controller:
   type: cartesian_impedance_controller/CartesianImpedanceController
   joints:                               # Joints to control
@@ -20,7 +24,7 @@ CartesianImpedance_trajectory_controller:
     stiffness: 0.1                      # Cartesian and nullspace stiffness
     wrench: 0.1                         # Commanded torque
   verbosity:
-    verbose_print: true                # Enables additional prints
+    verbose_print: false                # Enables additional prints
     state_msgs: true                    # Messages of controller state
     tf_frames: true                     # Extra tf frames
 
