### 2023-11-03 97cbfe1
New minimal example, now with 7 joints
