+++ res/config/example_controller.yaml
@@ -7,6 +7,7 @@ CartesianImpedance_trajectory_controller:
     - joint4
     - joint5
     - joint6
+    - joint7
   end_effector: tool0                   # Link to control arm in
   update_frequency: 500                 # Controller update frequency in Hz
   dynamic_reconfigure: true             # Starts dynamic reconfigure server
@@ -38,3 +39,4 @@ hardware_interface:
     - joint4
     - joint5
     - joint6
+    - joint7
