### 2023-11-03 3610038
Testing: Introduces new ROS functionality tests
