### 2023-11-07 4c92f24
Chore: Removes duplicate launch, yaml and urdf
