### 2023-11-07 f3ecc6a
Chore: Adds yaml-cpp as dependency and use catkin_make
