### 2023-11-07 9e7dd16
Chore: Adds rosdep to install script
