### 2023-11-07 #5194 @matthias-mayr
Thanks for your patience. We are finally there and all the changes are merged.

To sum up:
* we added unit tests for both the base library and the ROS implementation
  * the ROS environment also spawns the controller in simulation, sends a goal and checks that it moves as expected
* added an easy to launch startup example that does not need external dependencies
* added dependencies that are not present in a minimal ROS environment

FYI, the example robot as it's spawned in Gazebo:
![image](https://github.com/openjournals/joss-reviews/assets/64466074/d300d247-cc3b-4083-a144-39db22d913c4)

I think with that everything is addressed. Let us know if there is anything we can do.
