### 2023-11-07 79de5c7
Fix: Melodic python2-compatibility for install script
