### 2023-12-27 10e33ac
Chore: Updates task variations paper with DOI & pages
