+++ res/paper.bib
@@ -48,13 +48,14 @@
 
 @INPROCEEDINGS{ahmad23generalization,
   author={Ahmad, Faseeh and Mayr, Matthias and Krueger, Volker},
-  booktitle={IEEE/RSJ International Conference on Intelligent Robots and Systems (IROS)},
-  title={Learning to Adapt the Parameters of Behavior Trees and Motion Generators to Task Variations},
+  booktitle={2023 IEEE/RSJ International Conference on Intelligent Robots and Systems (IROS)},
+  title={Learning to Adapt the Parameters of Behavior Trees and Motion Generators (BTMGs) to Task Variations},
   year={2023},
   volume={},
   number={},
-  pages={},
-  doi={}}
+  pages={10133-10140},
+  doi={10.1109/IROS55552.2023.10341636}}
+
 
 @book{springer:2016,
   title={Springer Handbook of Robotics},
