### 2023-12-27 #5194 @matthias-mayr
One of the papers that uses this got formally published and I updated the reference.

Other than the reviewers reaction it looks to me like nothing is in the way of a publication and it would be nice to be able to remove this from the back of my head. Bonus points for a DOI before 03.01. if that could be feasible, so I can enter it into the final version of my PhD thesis.
