### 2024-01-15 #5194 @matthias-mayr
Here it is:
- Version: 1.1.1
- Zenodo: https://zenodo.org/records/10513692
- GH release: https://github.com/matthias-mayr/Cartesian-Impedance-Controller/releases/tag/v1.1.1
- DOI: 10.5281/zenodo.10513691

Let me know if there's anything else that I can do.
