### 2024-01-22 92d019f
Edits to JOSS paper
Ensures agreement with plural noun at beginning, removes italics on e.g. and i.e., removes extraneous parentheses around equations, and adds Oxford comma.
