+++ res/paper.md
@@ -30,24 +30,24 @@ A Cartesian impedance controller is a type of control strategy that is used in r
 
 In this package, we provide a C++ implementation of a controller that allows collaborative robots:
 
-1. To achieve compliance in its Cartesian task-frame coordinates.
-2. To allow for joint compliance in the nullspace of its task-frame coordinates.
-3. To be able to apply desired forces and torques to the environment of the robot, *e.g.*, for direct force control.
+1. To achieve compliance in Cartesian task-frame coordinates;
+2. To allow for joint compliance in the nullspace of task-frame coordinates; and
+3. To be able to apply desired forces and torques to their environments, e.g., for direct force control.
 
 This package can be used in any torque-controlled robotic manipulator. Its implementation in Robot Operating System (ROS) integrates it into `ros_control` [@ros_control] and can automatically utilize the URDF description of the robot's geometry.
 
 # Statement of Need
-Modern robotics is moving more and more past the traditional robot systems that have hard-coded paths and stiff manipulators. Many use-cases require the robots to work in semi-structured environments. These environments impose uncertainties that could cause collisions. Furthermore, many advanced assembly, manufacturing and household scenarios such as insertions or wiping motions require the robot to excert a controlled force on the environment. Finally, the robot workspace is becoming increasingly shared with human workers in order to leverage both agents and allow them to complement each other.
+Modern robotics is moving more and more past the traditional robot systems that have hard-coded paths and stiff manipulators. Many use cases require the robots to work in semi-structured environments. These environments impose uncertainties that could cause collisions. Furthermore, many advanced assembly, manufacturing, and household scenarios such as insertions or wiping motions require the robot to excert a controlled force on the environment. Finally, the robot workspace is becoming increasingly shared with human workers to leverage both agents and allow them to complement each other.
 
 An implementation of compliant control for robotic manipulators is an attractive solution for robots in contact-rich environments. To cover the wide variety of tasks and scenarios, we think that it needs to fulfill the following criteria:
 
 1. Dynamically adapt the end-effector reference point.
-2. Dynamically adjust the robot's impedance (*i.e.*, its ability to resist or comply with external forces).
-3. Apply commanded forces and torques (*i.e.*, a wrench) with the end-effector of the robot.
+2. Dynamically adjust the robot's impedance (i.e., its ability to resist or comply with external forces).
+3. Apply commanded forces and torques (i.e., a wrench) with the end-effector of the robot.
 4. Command a joint configuration and apply it in the nullspace of the Cartesian robotic task.
 5. Execute joint-space trajectories.
 
-A complete implementation with respect to items 1-5 above of compliance for torque-commanded robotic manipulators is not available, and the existing solutions `franka_ros` [@franka_ros] and `libfranka` [@libfranka] as well as the `KUKA` FRI Cartesian impedance controller can only be used for a single type of robotic manipulator:
[…]
