### 2024-01-22 #5194 @kyleniemeyer
Hi @matthias-mayr, a few things to address before publishing. Could you review and merge the edits I made in https://github.com/matthias-mayr/Cartesian-Impedance-Controller/pull/18; also, please make sure the acronym URDF is defined the paper (since it is only used once, it should just be spelled out).

Also, it looks like an extraneous contributor ended up on the Zenodo archive. Can you either add them to the paper author list, or remove?
