### 2024-01-23 #5194 @matthias-mayr
Thanks a lot for taking the time to go through the paper. I added a spelled-out version of URDF, but still kept the abbreviation since there might be people familiar with the acronym, but not the spelled-out version.
Having added that, I accepted your PR

I also removed the extra author on Zenodo from there since he only changed two values in a single line.

Let me know if there's anything else I can do.
